+++ test/hci_rhf.jl
@@ -32,26 +32,25 @@ using ElemCo
   @test abs(E_hci + energies["HCI-correction"] - E_HCI_PT2_test) < epsilon
 
   
-  @set hci epsilon=1.e-5
-      
-  energies = @hci
-
+  energies = @hci begin
+    hci(epsilon=1.e-5)
+  end
   E_hci_tight = energies["HCI"]
 
   println("HCI Energy (tight selection, ε=1e-5): $E_hci_tight")
   @test abs(E_hci_tight - E_HCI_tight_test) < epsilon
 
-  @set hci epsilon=1.e-3
-
-  energies = @hci
-      
+  energies = @hci begin
+    hci(epsilon=1.e-3)
+  end
   E_hci_loose = energies["HCI"]
       
   println("HCI Energy (loose selection, ε=1e-3): $E_hci_loose")
   @test abs(E_hci_loose - E_HCI_loose_test) < epsilon
   
-  @set hci epsilon=5.e-4 nstates=3
-  energies = @hci
+  energies = @hci begin
+    @set hci epsilon=5.e-4 nstates=3
+  end
 
   println("HCI Multi-root energies: ", energies["HCI"], ", ", energies["ω1"], ", ", energies["ω2"])
   @test abs(energies["HCI"] - E_HCI_ms_test) < epsilon